Refactor FontTemplateInfo away
Rather than getting the webrender_api::FontKey in one place and the
webrender_api::FontInstanceKey in another, we can do them both when we
create the Font instance.

This also creates a slightly better separation between the code which is
doing font selection (which shouldn't need to know about WebRender)
and the rest.